fastapi/requirements.txt
fastapi
uvicorn

langgraph
langchain
langchain-groq
asyncio
pydantic
aiosmtplib
email-validator
celery
kombu
billiard
amqp
python-jose
boto3
python-dotenv
requests
reportlab
matplotlib
pandas
jinja2

duckduckgo-search

beautifulsoup4
